requirements.txt
wakeonlan
python-dotenv
python-telegram-bot
